requirements.txt
mfrc522==0.0.7
RPi.GPIO==0.7.1
spi==0.2.0
spidev==3.6
